requirements.txt
streamlit==1.32.0
langchain-huggingface
langchain-groq
langchain-text-splitters
qdrant-client
python-dotenv
sentence-transformers
groq